Refactoring before making RSSM and measurement probe models
