--- src/e2c.py
@@ -1,492 +1,0 @@
-"""
-e2c.py
-
-Embed to Control model architecture made with PyTorch.
-
-Authors: Jared Berry, Ayush Gaggar
-"""
-import torch
-from torch import nn
-from pathlib import Path
-import numpy as np
-
-from src.encode import ConvEncoder, ConvDecoder, ChannelUncertaintyConvDecoder, ScalarUncertaintyConvDecoder
-
-# Get paths relative to the project root
-PROJECT_ROOT = Path(__file__).parent.parent
-DATA_PATH = PROJECT_ROOT / "data"
-CONFIG_PATH = PROJECT_ROOT / "config"
-
-class E2CDataset(torch.utils.data.Dataset):
-    """
-    An E2C Dataset consists of a current image, a future image, and a control input.
-    """
-    def __init__(self, config):
-        # Load raw dataset
-        env_name = config['train']['dataset'].split('_')[0]
-        dataset_dir = DATA_PATH / env_name
-        data = torch.load(dataset_dir / f"{config['train']['dataset']}.pt")
-        self.X = data["prev_images"].permute(0, 1, 4, 2, 3)  # Shape: [num_samples, past_length, C, H, W]
-        self.X_next = data["next_images"].permute(0, 3, 1, 2)  # Shape: [num_samples, C, H, W], since pred_length=1
-        self.img_shape = [self.X[0, 0].shape[0] * config['trans']['past_length'], *self.X[0, 0].shape[1:]]
-        if len(self.X.shape) == 5:
-            self.past_length = self.X.shape[1]
-            assert self.past_length == config['trans']['past_length'], f"past_length={config['trans']['past_length']} in config file, but dataset has past_length={self.past_length}"
-        else:
-            self.past_length = 1
-        self.pred_length = config['trans']['pred_length']
-        control = data["actions"]
-
-        # Normalize actions
-        # if control.max() - control.min() > 0:
-        #     control = (control - control.min()) / (control.max() - control.min())
-        self.U = control
-
-        # Reshape for learning
-        # flat = img.reshape(-1, img.shape[2], img.shape[3], img.shape[4]).permute(0, 3, 1, 2)   # Shape: [batch*seq_len, C, H, W]
-        # X = flat[:-1]   # Shape: [(batch*seq_len) - 1, C, H, W]
-        # X_next = flat[1:]
-        # U = control.reshape(-1, control.shape[-1])[:-1]
-        # Create windowed samples for past_length and pred_length
-        """H, W = self.img_shape[1:]
-        num_batches = img.shape[0]
-        seq_len = img.shape[1]
-        x_windows = []
-        x_next_windows = []
-        u_list = []
-        window_size = self.past_length + self.pred_length
-        for b in range(num_batches):
-            for t in range(seq_len - window_size + 1):
-                # Get windows
-                x_windows.append(img[b, t:t+self.past_length].reshape(-1, H, W))
-                x_next_windows.append(img[b, t+self.past_length:t+self.past_length+self.pred_length].reshape(-1, H, W))
-                u_list.append(control[b, t+self.past_length])          
-
-        self.X = torch.stack(x_windows)
-        self.X_next = torch.stack(x_next_windows)
-        self.U = torch.stack(u_list)"""
-
-    def __len__(self):
-        return len(self.X)
-
-    def __getitem__(self, idx):
-        return self.X[idx], self.X_next[idx], self.U[idx]
-  
-class E2CLoss(nn.Module):
-    """
-    E2C loss, made with PyTorch.
-    """
-    def __init__(self, num_epochs, loss_params):
-        super().__init__()
-        self.num_epochs = num_epochs
-        self.recon_mult = loss_params['recon_mult']
-        self.beta = loss_params['beta']
-        self.lam = loss_params['lambda']
-        self.anneal_mode = loss_params['kld_anneal_mode']
-
-    def kld_anneal(self, epoch):
-        if self.anneal_mode == 'const':
-            mult = self.beta
-        elif self.anneal_mode == 'linear':
-            mult = self.beta*((epoch + 1)/self.num_epochs)
-        else:
-            raise NotImplementedError(f"Annealing mode {self.anneal_mode} not supported!")
-
-        return mult
-
-    def forward(self, tr, epoch):
-        # Reconstruction loss
-        recon = self.recon_mult*nn.functional.mse_loss(tr['x'], tr['x_recon'], reduction='mean')
-        recon_next = self.recon_mult*nn.functional.mse_loss(tr['x_next'], tr['x_pred'], reduction='mean')
-
-        # Encoding KL Divergence
-        log_var, mu = tr['log_var'], tr['mu']
-        kld = self.kld_anneal(epoch)*(-0.5 * torch.sum(1 + log_var - mu.pow(2) - log_var.exp(), dim=-1).mean())
-
-        # Transition model KLD
-        # https://stats.stackexchange.com/questions/7440/kl-divergence-between-two-univariate-gaussians
-        log_var_pred, mu_pred = tr['log_var_pred'], tr['mu_pred']
-        log_var_next, mu_next = tr['log_var_next'], tr['mu_next']
-        var_next, var_pred = torch.exp(log_var_next), torch.exp(log_var_pred)
-
-        # Times 0.5 because formula uses standard dev
-        # kld_trans_vec = 0.5*(log_var_pred - log_var_next) \
-        #                    + ((var_next + (mu_next - mu_pred)**2) / (var_pred + 1e-8)) \
-        #                    - 1.0
-        # kld_trans = 0.5 * torch.sum(kld_trans_vec)
-
-        # v, r = tr['v'], tr['r']
-        # dot = torch.sum(v * r, dim=1)
-        # dot = torch.clamp(dot, min=-0.99)  # Ensure log(1 + dot) > 0
-        # sum_term = torch.sum(log_var_pred - log_var_next, axis=1)
-        # log_term = torch.log(1 + dot)
-        # kld_trans_vec = 2*(sum_term - log_term)
-        # kld_trans = self.lam*torch.sum(kld_trans_vec)
-
-        # def kl_diag_gaussians(mu_p, log_var_p, mu_q, log_var_q):
-        #     var_p = torch.exp(log_var_p)
-        #     var_q = torch.exp(log_var_q)
-
-        #     return 0.5 * torch.sum(
-        #         log_var_q - log_var_p
-        #         + (var_p + (mu_p - mu_q).pow(2)) / var_q
-        #         - 1,
-        #         dim=-1
-        #     )
-        # kld_trans = self.lam * kl_diag_gaussians(
-        #     mu_pred, log_var_pred,
-        #     mu_next, log_var_next
-        # ).mean()
-
-        z_pred = tr['z_pred']
-        kld_trans = self.lam*(-0.5 * torch.sum(1 + log_var_next - (z_pred - mu_next).pow(2) - log_var_next.exp(), dim=-1).mean())
-
-        loss = recon + recon_next + kld + kld_trans
-        if torch.isnan(loss):
-            breakpoint()
-
-        # Make return dictionary for loss values
-        loss_return = {
-            r"$x$ Reconstruction Loss": recon.detach().cpu().item(),
-            r"$x_{next}$ Reconstruction Loss": recon_next.detach().cpu().item(),
-            "KLD": kld.detach().cpu().item(),
-            "Transition KLD": kld_trans.detach().cpu().item()
-        }
-        return loss, loss_return
-    
-class UncertaintyE2CLoss(E2CLoss):
-    """
-    E2C loss with full image prediction uncertainty estimates, with reconstruction losses as negative loss likelihood.
-    """
-    def __init__(self, num_epochs, loss_params):
-        super().__init__(num_epochs, loss_params)
-
-    def expand_uncertainty(self, logvar, target):
-        """
-        Ensure uncertainty (log variance) output from decoder is of correct shape to be used by 
-        gaussian_nll loss function in PyTorch API.
-
-        Args:
-            logvar: uncertainty tensor (any of:
-                    [B], [B,1], [B,1,1,1], [B,K,H,W], [B,C,H,W])
-                    where C % k == 0
-            target: image tensor [B,C,H,W]
-        """
-        B, C, H, W = target.shape
-
-        # Ensure 4D
-        while logvar.dim() < 4:
-            logvar = logvar.unsqueeze(-1)
-
-        # [B,1,1,1] -> [B,C,H,W]
-        if logvar.shape[1] == 1:
-            logvar = logvar.expand(B, C, H, W)
-
-        # [B,K,H,W] where K != C
-        elif logvar.shape[1] != C:
-            if C % logvar.shape[1] != 0:
-                raise ValueError(
-                    f"Cannot expand uncertainty with {logvar.shape[1]} channels "
-                    f"to match image with {C} channels"
-                )
-            
-            # Tile uncertainty to be shape compatible
-            repeat_factor = C // logvar.shape[1]
-            logvar = logvar.repeat(1, repeat_factor, 1, 1)
-
-        return logvar
-
-    def forward(self, tr, epoch):
-        # Ensure uncertainty shape is compatible with loss function
-        x_recon_uncertainty = self.expand_uncertainty(tr['x_recon_uncertainty'], tr['x_recon'])
-        x_pred_recon_uncertainty = self.expand_uncertainty(tr['x_pred_recon_uncertainty'], tr['x_pred'])
-        recon_nll = nn.functional.gaussian_nll_loss(
-            tr['x_recon'],
-            tr['x'], 
-            torch.exp(x_recon_uncertainty) + 1e-6,    # Train decoder uncertainty outputs to be log variance
-            reduction='mean'
-        )
-        pred_nll = nn.functional.gaussian_nll_loss(
-            tr['x_pred'],
-            tr['x_next'], 
-            torch.exp(x_pred_recon_uncertainty) + 1e-6,
-            reduction='mean'
-        )
-        # recon_nll = nn.functional.gaussian_nll_loss(
-        #     tr['x_recon'],
-        #     tr['x'], 
-        #     torch.exp(tr['x_recon_uncertainty'].view(-1, 1, 1, 1).expand_as(tr['x_recon'])) + 1e-6,    # Train decoder uncertainty outputs to be log variance
-        #     reduction='mean'
-        # )
-        # pred_nll = nn.functional.gaussian_nll_loss(
-        #     tr['x_pred'],
-        #     tr['x_next'], 
-        #     torch.exp(tr['x_pred_recon_uncertainty'].view(-1, 1, 1, 1).expand_as(tr['x_pred'])) + 1e-6,
-        #     reduction='mean'
-        # )
-        # recon_logvar = tr['x_recon_uncertainty'].expand_as(tr['x_recon']) + 1e-6
-        # recon_var = torch.exp(recon_logvar ** 2) + 1e-6
-        # recon_nll = -((tr['x'] - tr['x_recon']) ** 2) / (2 * recon_var) - recon_logvar - np.log(np.sqrt(2 * np.pi))
-        # recon_nll = recon_nll.mean()
-
-        # pred_logvar = tr['x_pred_recon_uncertainty'].expand_as(tr['x_pred']) + 1e-6
-        # pred_var = torch.exp(pred_logvar ** 2) + 1e-6
-        # pred_nll = -((tr['x_next'] - tr['x_pred']) ** 2) / (2 * pred_var) - pred_logvar - np.log(np.sqrt(2 * np.pi))
-        # pred_nll = pred_nll.mean()
-
-        recon = recon_nll + self.recon_mult*nn.functional.mse_loss(tr['x'], tr['x_recon'], reduction='mean')
-        recon_next = pred_nll + self.recon_mult*nn.functional.mse_loss(tr['x_next'], tr['x_pred'], reduction='mean')
-
-        # Encoding KL Divergence
-        log_var, mu = tr['log_var'], tr['mu']
-        kld = self.kld_anneal(epoch)*(-0.5 * torch.sum(1 + log_var - mu.pow(2) - log_var.exp(), dim=-1).mean())
-
-        # Transition model KLD
-        # https://stats.stackexchange.com/questions/7440/kl-divergence-between-two-univariate-gaussians
-        log_var_pred, mu_pred = tr['log_var_pred'], tr['mu_pred']
-        log_var_next, mu_next = tr['log_var_next'], tr['mu_next']
-        z_pred = tr['z_pred']
-        kld_trans = self.lam*(-0.5 * torch.sum(1 + log_var_next - (z_pred - mu_next).pow(2) - log_var_next.exp(), dim=-1).mean())
-
-        loss = recon + recon_next + kld + kld_trans
-        if torch.isnan(loss):
-            breakpoint()
-
-        # Make return dictionary for loss values
-        loss_return = {
-            r"$x$ Reconstruction Loss": recon.detach().cpu().item(),
-            r"$x_{next}$ Reconstruction Loss": recon_next.detach().cpu().item(),
-            r"$x$ NLL Reconstruction Loss": recon_nll.detach().cpu().item(),
-            r"$x_{next}$ NLL Reconstruction Loss": pred_nll.detach().cpu().item(),
-            "KLD": kld.detach().cpu().item(),
-            "Transition KLD": kld_trans.detach().cpu().item()
-        }
-        return loss, loss_return
-
-class ConvE2C(nn.Module):
-    """
-    An E2C model with convolutional encoder-decoder.
-
-    Args:
-        enc_latent_size: Latent dimension of encoder
-        latent_size: E2C latent dimension
-        control_size: Dimension of control vector
-        past_length: Length of training observation history
-        pred_length: Prediction horizon length
-        conv_params: Dictionary containing CNN params for encoder/decoder
-        device: Torch device object
-        uncertainty_output: Whether to use ChannelUncertainty decoder
-    """
-    def __init__(self, enc_latent_size, latent_size, control_size, past_length, pred_length, conv_params, device, output_uncertainty=False):
-        super().__init__()
-        self.device = device
-        self.output_uncertainty = output_uncertainty
-
-        # Set number of hidden units
-        self.enc_latent_size = enc_latent_size
-        self.latent_size = latent_size
-        self.control_size = control_size
-        self.past_length = past_length
-        self.pred_length = pred_length
-
-        # Dummy zero control vector
-        self.dummy_u = torch.zeros((1, self.control_size)).to(self.device)
-
-        # Encoder and decoder
-        in_channels = conv_params['out_image_shape'][0]
-        self.encoder = ConvEncoder(enc_latent_size, in_channels, conv_params)
-        if self.output_uncertainty:
-            self.decoder = ChannelUncertaintyConvDecoder(latent_size, conv_params, self.encoder.out_dim_flat, self.encoder.out_shape)
-        else:
-            self.decoder = ConvDecoder(latent_size, conv_params, self.encoder.out_dim_flat, self.encoder.out_shape)
-
-        # VAE
-        self.mu = nn.Linear(self.enc_latent_size, self.latent_size)
-        self.log_var = nn.Linear(self.enc_latent_size, self.latent_size)
-
-        # Locally linear transition model
-        self.fc_trans = nn.Sequential(
-            nn.Linear(self.latent_size, 200),
-            nn.ReLU(),
-            nn.Linear(200, 200),
-            nn.ReLU(),
-            nn.Linear(200, 32)
-        )
-        self.v = nn.Linear(32, self.latent_size)  # A jacobian
-        self.r = nn.Linear(32, self.latent_size)
-        self.B = nn.Linear(32, self.latent_size*self.control_size)
-        self.o = nn.Sequential(
-            nn.Linear(32, self.latent_size),
-            nn.Tanh()
-        )
-        self.H = nn.Sequential(
-            nn.Linear(32, self.latent_size),
-            nn.Softplus() # Ensure positive covariances
-        )
-
-    def reparameterize(self, mu, log_var):
-        # Get standard deviation from log variance
-        std = torch.exp(0.5 * log_var)
-        std = torch.clamp(std, min=1e-5, max=1e5) # Prevent std from being too small
-
-        # Generate random noise epsilon of same shape std
-        eps = torch.randn_like(std)
-
-        # Return reparameterized sample
-        return mu + eps * std
-
-    def transition(self, z, mu, log_var, u):
-        # Pass through transition network
-        trn = self.fc_trans(z)
-
-        # Get Jacobians and linear model parameters
-        v = self.v(trn).unsqueeze(-1)                                                    #   [batch, z, 1]
-        r = self.r(trn).unsqueeze(-1)                                                    #   [batch, z, 1]
-        A = torch.eye(self.latent_size, device=self.device).repeat(z.size(0), 1, 1) \
-            + torch.bmm(v, r.transpose(1, 2))                                            #   [batch, z, z]
-        B = self.B(trn).reshape((-1, self.latent_size, self.control_size))               #   [batch, z, u]
-        o = self.o(trn).unsqueeze(-1)                                                    #   [batch, z, 1]
-        H = torch.diag_embed(self.H(trn))                                                #   [batch, z, z]
-
-        # Use linear model and reparam to create posterior distribution
-        mu_pred = torch.bmm(A, mu.unsqueeze(-1)) + torch.bmm(B, u.unsqueeze(-1)) + o     #   [batch, z, 1]
-        mu_pred = mu_pred.squeeze(-1)                                                    #   [batch, z]
-        sigma = torch.diag_embed(torch.exp(log_var))                                     #   [batch, z, z]
-        C = torch.bmm(A, torch.bmm(sigma, A.transpose(1, 2))) + H                        #   [batch, z, z]
-        log_var_pred = torch.log(torch.diagonal(C, dim1=-2, dim2=-1) + 1e-8)             #   [batch, z]
-        z_pred = self.reparameterize(mu_pred, log_var_pred)                              #   [batch, z]
-
-        return mu_pred, log_var_pred, z_pred
-
-    def forward(self, x, x_next, u):
-        # Encode current and next state
-        enc_out = self.encoder(x)
-        enc_out_next = self.encoder(x_next)
-
-        # Get latent variable
-        mu = self.mu(enc_out)
-        log_var = self.log_var(enc_out)
-        z = self.reparameterize(mu, log_var)
-
-        # Get next latent variable
-        mu_next = self.mu(enc_out_next)
-        log_var_next = self.log_var(enc_out_next)
-        z_next = self.reparameterize(mu_next, log_var_next)
-
-        # Transition model
-        mu_pred, log_var_pred, z_pred = self.transition(z, mu, log_var, u)
-
-        # Get reconstruction and prediction
-        if self.output_uncertainty:
-            x_recon, x_recon_uncertainty = self.decoder(z)
-            x_next_recon, x_next_recon_uncertainty = self.decoder(z_next)
-            x_pred, x_pred_recon_uncertainty = self.decoder(z_pred)
-            train_return = {
-                'x_recon': x_recon,
-                'mu': mu,
-                'log_var': log_var,
-                'x_next_recon': x_next_recon,
-                'mu_next': mu_next,
-                'log_var_next': log_var_next,
-                'x_pred': x_pred,
-                'z_pred': z_pred,
-                'mu_pred': mu_pred,
-                'log_var_pred': log_var_pred,
-                'x_recon_uncertainty': x_recon_uncertainty,
-                'x_next_recon_uncertainty': x_next_recon_uncertainty,
-                'x_pred_recon_uncertainty': x_pred_recon_uncertainty
-            }
-        else:
-            x_recon = self.decoder(z)
-            x_next_recon = self.decoder(z_next)
-            x_pred = self.decoder(z_pred)
-            train_return = {
-                'x_recon': x_recon,
-                'mu': mu,
-                'log_var': log_var,
-                'x_next_recon': x_next_recon,
-                'mu_next': mu_next,
-                'log_var_next': log_var_next,
-                'x_pred': x_pred,
-                'z_pred': z_pred,
-                'mu_pred': mu_pred,
-                'log_var_pred': log_var_pred,
-            }
-
-        return train_return
-
-    def reconstruct(self, x_traj):
-        """
-        Reconstruct an entire trajectory to test encoder/decoder
-        """
-        with torch.no_grad():
-            frames = []
-            for x in x_traj:
-                # Encode current state
-                encoded = self.encoder(x.unsqueeze(0))
-                flattened = encoded.view(encoded.size(0), -1)
-
-                # Get latent variable
-                mu = self.mu(flattened)
-                log_var = self.log_var(flattened)
-                z = self.reparameterize(mu, log_var)
-                frames.append(self.decoder(z))
-
-            return torch.concat(frames, dim=0).squeeze(0).to('cpu').permute(0, 2, 3, 1)
-
-    def sample_traj(self, x0, seq_len):
-        """
-        Sample an entire trajectory, starting from an initial condition
-        """
-        with torch.no_grad():
-            # Encode current state
-            encoded = self.encoder(x0.unsqueeze(0).to(self.device))
-            flattened = encoded.view(encoded.size(0), -1)
-
-            # Get latent variable
-            mu = self.mu(flattened)
-            log_var = self.log_var(flattened)
-            z = self.reparameterize(mu, log_var)
-
-            # Transition model
-            frames = []
-            frames.append(self.decoder(z))
-            for t in range(seq_len):
-                mu, log_var, z, _, _ = self.transition(z, mu, log_var, self.dummy_u)
-                frames.append(self.decoder(z))
-
-            return torch.concat(frames, dim=0).squeeze(0).to('cpu').permute(0, 2, 3, 1)
-        
-    def sample(self, x, u, return_all=False):
-        """
-        Predict the next image in a sequence
-        """
-        with torch.no_grad():
-            sample_return = {}
-            # Encode current state
-            encoded = self.encoder(x.to(self.device))
-            flattened = encoded.view(encoded.size(0), -1)
-
-            # Get latent variable
-            mu = self.mu(flattened)
-            log_var = self.log_var(flattened)
-            z = self.reparameterize(mu, log_var)
-
-            # Predict transition and decode current pred and next pred
-            mu_pred, log_var_pred, z_pred = self.transition(z, mu, log_var, u)
-            sample_return['mu_pred'] = mu_pred
-            sample_return['log_var_pred'] = log_var_pred
-            sample_return['z_pred'] = z_pred
-            if self.output_uncertainty:
-                x_recon, x_recon_uncertainty = self.decoder(z)
-                x_pred, x_pred_recon_uncertainty = self.decoder(z_pred)
-                sample_return['x_recon_uncertainty'] = x_recon_uncertainty
-                sample_return['x_pred_recon_uncertainty'] = x_pred_recon_uncertainty
-            else:
-                x_recon = self.decoder(z)
-                x_pred = self.decoder(z_pred)
-
-            if return_all and self.output_uncertainty:
-                return x_recon.squeeze(0), x_pred.squeeze(0), sample_return
-            else:
-                return x_recon.squeeze(0), x_pred.squeeze(0)
